+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - thermal_conversion parameter to track the conversion of the thermal_power array to cosmological units
+- cosmology object support in UnitParameters
 ### Changed
+- Various typos
+- Docstring Formatting
 ### Fixed
 - definition of acceptable units for data_array and noise_array. Previously had units.K not units.K * units.sr * units.Hz
 - definition of units for power_array and noise_power when data is uncalibrated